@@ -32,8 +32,8 @@ class SchemaTypeBuffer extends SchemaType {
     const value = super.cast(value_, data);
 
     if (value == null || Buffer.isBuffer(value)) return value;
-    if (typeof value === 'string') return new Buffer(value, this.options.encoding);
-    if (Array.isArray(value)) return new Buffer(value);
+    if (typeof value === 'string') return Buffer.from(value, this.options.encoding);
+    if (Array.isArray(value)) return Buffer.from(value);
   }
 
   /**
@@ -76,7 +76,7 @@ class SchemaTypeBuffer extends SchemaType {
    * @return {Boolean}
    */
   parse(value, data) {
-    return value ? new Buffer(value, this.options.encoding) : value;
+    return value ? Buffer.from(value, this.options.encoding) : value;
   }
 
   /**
