@@ -1,6 +1,6 @@
 'use strict';
 
-const should = require('chai').should();
+const should = require('chai').use(require('chai-as-promised')).should();
 const sortBy = require('lodash/sortBy');
 const Promise = require('bluebird');
 const sinon = require('sinon');
@@ -84,7 +84,7 @@ describe('Model', () => {
       age: 20
     }).then(data => {
       User.findById(data._id).should.exist;
-      User.length.should.eql(1);
+      User.should.to.have.length(1);
       listener.calledOnce.should.be.true;
       return data;
     }).then(data => User.removeById(data._id));
@@ -197,7 +197,7 @@ describe('Model', () => {
       const user = User.findById(id);
 
       user.age.should.eql(2);
-      User.length.should.eql(1);
+      User.should.to.have.length(1);
 
       return User.removeById(id);
     });
@@ -470,7 +470,7 @@ describe('Model', () => {
   });
 
   it('count()', () => {
-    Post.length.should.eql(Post.count());
+    Post.should.to.have.length(Post.count());
   });
 
   it('forEach()', () => {
@@ -1217,7 +1217,7 @@ describe('Model', () => {
       {_id: 'B', bool: 0}
     ]);
 
-    Test.length.should.eql(2);
+    Test.should.to.have.length(2);
 
     Test.toArray().should.eql([
       Test.new({_id: 'A', bool: true}),
