@@ -1,6 +1,6 @@
 'use strict';
 
-const should = require('chai').should(); // eslint-disable-line
+require('chai').should();
 const ValidationError = require('../../../lib/error/validation');
 
 describe('SchemaTypeEnum', () => {
@@ -11,24 +11,12 @@ describe('SchemaTypeEnum', () => {
 
     type.validate('foo').should.eql('foo');
 
-    try {
-      type.validate('wat');
-    } catch (err) {
-      err.should.be
-        .instanceOf(ValidationError)
-        .property('message', 'The value must be one of foo, bar, baz');
-    }
+    (() => type.validate('wat')).should.to.throw(ValidationError, 'The value must be one of foo, bar, baz');
   });
 
   it('validate() - required', () => {
     const type = new SchemaTypeEnum('test', {required: true});
 
-    try {
-      type.validate();
-    } catch (err) {
-      err.should.be
-        .instanceOf(ValidationError)
-        .property('message', '`test` is required!');
-    }
+    type.validate.bind(type).should.to.throw(ValidationError, '`test` is required!');
   });
 });
